requirements.txt
pytest
requests
PyYAML
PyMySQL
allure-pytest